@@ -1,8 +1,8 @@
 package demo.app.controllers;
 
 import com.microsoft.applicationinsights.TelemetryClient;
-import javax.servlet.RequestDispatcher;
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.RequestDispatcher;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.web.servlet.error.ErrorController;
 import org.springframework.http.HttpStatus;
@@ -23,7 +23,7 @@ public class CustomErrorController implements ErrorController {
   @GetMapping("/error")
   public String handleError(Model model, HttpServletRequest request) {
     int status = (Integer) request.getAttribute(RequestDispatcher.ERROR_STATUS_CODE);
-    Exception exception = (Exception) request.getAttribute("javax.servlet.error.exception");
+    Exception exception = (Exception) request.getAttribute("jakarta.servlet.error.exception");
     String message = "";
     System.out.println("### telemetryClient = " + telemetryClient);
     if (exception != null) {    
@@ -37,10 +37,5 @@ public class CustomErrorController implements ErrorController {
     model.addAttribute("message", message);
     model.addAttribute("status", status);
     return "error";
-  }
-
-  @Override
-  public String getErrorPath() {
-    return "/error";
   }
 }
